requirements.txt
beautifulsoup4
joblib
numpy
pandas
pyinstaller
pywebview
requests
scikit-learn
streamlit
